requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
flake8==6.0.0
idna==3.4
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
